requirements.txt
aiodns==1.1.1
aiohttp==3.6.2
aiohttp-socks==0.2.2
async-timeout==3.0.1
attrs==19.3.0
autobahn==19.11.1
Automat==0.8.0
backtrader==1.9.74.123
bitstring==3.1.6
ccxt==1.19.50
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
constantly==15.1.0
cryptography==2.8
Cython==0.29.14
dateparser==0.7.2
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
multidict==4.5.2
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycares==3.0.0
pycparser==2.19
PyHamcrest==1.9.0
pyOpenSSL==19.0.0
python-binance==0.7.4
python-dateutil==2.8.1
pytz==2019.3
regex==2019.11.1
requests==2.22.0
service-identity==18.1.0
six==1.12.0
Twisted==19.10.0
txaio==18.8.1
tzlocal==2.0.0
ujson==1.35
urllib3==1.25.7
websocket-client==0.56.0
yarl==1.1.0
zope.interface==4.7.1
